Bug 679832: Explicitly cast to a pointer-sized integer, when converting between 32-bit ints and void* pointers in nsCheapSets.h. r=bsmedberg
UltraBlame original commit: 196df669baba497e9d2aabdd1dcc33f120d1bca5

+++ xpcom/ds/nsCheapSets.h
@@ -175,7 +175,7 @@ private:
   
   void SetInt(PRInt32 aInt)
   {
-    mValOrHash = (void*)((aInt << 1) | 0x1);
+    mValOrHash = (void*)(intptr_t)((aInt << 1) | 0x1);
   }
   
   nsresult InitHash(nsInt32HashSet** aSet);
